refactor: create separate files for util methods and workspace verbs
