+++ rtwcli/rtwcli/setup.py
@@ -19,6 +19,7 @@ setup(
     name="rtwcli",
     version="0.2.0",
     packages=find_packages(exclude=["test"]),
+    install_requires=["docker", "questionary"],
     extras_require={
         "completion": ["argcomplete"],
     },
